Rename `pathfinding_service/config.py`
for consistency with `monitoring_service/constants.py` and
`raiden/constants.py`
