+++ src/pathfinding_service/constants.py
