+++ tests/pathfinding/test_model.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
 
-from pathfinding_service.config import MAX_AGE_OF_FEEDBACK_REQUESTS
+from pathfinding_service.constants import MAX_AGE_OF_FEEDBACK_REQUESTS
 from pathfinding_service.model.feedback import FeedbackToken
 from raiden.utils.typing import TokenNetworkAddress
